Cleanly restore text on clearing shallow marks
When removing the highlight divisions were being left. If "ba" was
searched in "foo bar baz" and then removed you were left with

    "ba"
    "r baz"

When the next highlight occurs with the string "bar" each text node is
surrounded thusly:

    <span>ba</span>
    <span>r</span>
    " baz"

When cleared this leads to:

    "ba"
    "r"
    " baz"

This pattern then continues to get worse in perpetuity.