Remove visibility filter from shallow search
By filtering for visibility as part of `findAndReplaceDomText` we could
ensure that clicking "Find" and "Find Prev" would always jump to a
visible element. Unfortunately, because we perform this operation in
memory, `:visible` is no longer a meaningful concept. This logic will
have to be deferred until "Find"/"Find Prev" is run.